+++ src/CxbxKrnl/EmuXapi.cpp
@@ -54,7 +54,7 @@ namespace xboxkrnl
 #include "EmuXTL.h"
 #include "EmuShared.h"
 #include "../Common/Win32/XBPortMapping.h"
-#include "HLEIntercept.h"
+#include "core/HLE/Intercept.hpp"
 #include "CxbxVSBC/CxbxVSBC.h"
 #include "Windef.h"
 #include <vector>
